fix: api playground loading state
