+++ packages/ui/app/src/css/globals.scss
@@ -7,7 +7,7 @@
 @import "./components";
 @import "../syntax-highlighting/FernSyntaxHighlighter";
 @import "../api-reference";
-@import "../playground/PlaygroundEndpoint";
+@import "../playground";
 @import "../components";
 @import "../mdx/components";
 @import "./utilities";
