+++ packages/ui/app/src/playground/PlaygroundDrawer.tsx
@@ -13,7 +13,6 @@ import { HEADER_HEIGHT_ATOM, useAtomEffect, useFlattenedApis, useSidebarNodes } 
 import {
     MAX_PLAYGROUND_HEIGHT_ATOM,
     PLAYGROUND_NODE_ID,
-    useHasPlayground,
     useIsPlaygroundOpen,
     usePlaygroundFormStateAtom,
     usePlaygroundNode,
@@ -22,14 +21,18 @@ import {
 import { IS_MOBILE_SCREEN_ATOM, MOBILE_SIDEBAR_ENABLED_ATOM, VIEWPORT_HEIGHT_ATOM } from "../atoms/viewport";
 import { FernErrorBoundary } from "../components/FernErrorBoundary";
 import { isEndpoint, isWebSocket, type ResolvedApiEndpointWithPackage } from "../resolver/types";
-import { PlaygroundEndpoint } from "./PlaygroundEndpoint";
-import { PlaygroundEndpointSelectorContent, flattenApiSection } from "./PlaygroundEndpointSelectorContent";
 import { PlaygroundWebSocket } from "./PlaygroundWebSocket";
 import { HorizontalSplitPane } from "./VerticalSplitPane";
+import { PlaygroundEndpoint } from "./endpoint/PlaygroundEndpoint";
+import { PlaygroundEndpointSelectorContent, flattenApiSection } from "./endpoint/PlaygroundEndpointSelectorContent";
+import { PlaygroundEndpointSkeleton } from "./endpoint/PlaygroundEndpointSkeleton";
 import { useResizeY } from "./useSplitPlane";
 
-export const PlaygroundDrawer = memo((): ReactElement | null => {
-    const hasPlayground = useHasPlayground();
+interface PlaygroundDrawerProps {
+    isLoading: boolean;
+}
+
+export const PlaygroundDrawer = memo(({ isLoading }: PlaygroundDrawerProps): ReactElement | null => {
     const selectionState = usePlaygroundNode();
     const apis = useFlattenedApis();
 
@@ -115,15 +118,13 @@ export const PlaygroundDrawer = memo((): ReactElement | null => {
 
     const setFormState = useSetAtom(usePlaygroundFormStateAtom(selectionState?.id ?? FernNavigation.NodeId("")));
 
-    if (!hasPlayground || apiGroups.length === 0) {
-        return null;
-    }
-
     const renderContent = () =>
         selectionState?.type === "endpoint" && matchedEndpoint != null ? (
             <PlaygroundEndpoint endpoint={matchedEndpoint} types={types} />
         ) : selectionState?.type === "webSocket" && matchedWebSocket != null ? (
             <PlaygroundWebSocket websocket={matchedWebSocket} types={types} />
+        ) : isLoading ? (
+            <PlaygroundEndpointSkeleton />
         ) : (
             <div className="size-full flex flex-col items-center justify-center">
                 <ArrowLeft className="size-8 mb-2 t-muted" />
