+++ packages/ui/app/src/playground/endpoint/PlaygroundCardSkeleton.tsx
@@ -1,0 +1,13 @@
+import clsx from "clsx";
+import { PropsWithChildren, ReactElement } from "react";
+
+export function PlaygroundCardSkeleton({
+    className,
+    children,
+}: PropsWithChildren<{ className?: string }>): ReactElement {
+    return (
+        <div className={clsx("rounded-xl bg-tag-default", className)}>
+            {children && <div className="contents invisible">{children}</div>}
+        </div>
+    );
+}
