+++ packages/ui/app/src/playground/endpoint/PlaygroundEndpoint.scss
