+++ packages/ui/app/src/playground/index.scss
@@ -1,0 +1,1 @@
+@import "./endpoint/PlaygroundEndpoint";
